Extract impl of serverless feeder
